+++ CaptchaSharp/Services/AntiCaptcha/Responses/Solutions/ImageCaptchaSolution.cs
@@ -7,7 +7,7 @@ namespace CaptchaSharp.Services.AntiCaptcha.Responses.Solutions
         public string Text { get; set; }
         public string Url { get; set; }
 
-        public CaptchaResponse ToCaptchaResponse(int id)
+        public CaptchaResponse ToCaptchaResponse(long id)
         {
             return new StringResponse
             {
